Move type annotations tests to their own class
